+++ CHANGES.md
@@ -4,6 +4,7 @@ JAMES Extensions Module Changes
 Version 1.1.1 (under development)
 ---------------------------------
 
+ - Added `NormalizedObjective` wrapper.
  - Various small improvements.
  
 Version 1.1 (11/07/2015)