+++ CHANGES.md
@@ -5,6 +5,7 @@ Version 1.1.1 (under development)
 ---------------------------------
 
  - Added `NormalizedObjective` wrapper.
+ - Added `toString` implementations to various components.
  - Various small improvements.
  
 Version 1.1 (11/07/2015)